feat: 주기적 가격 업데이트 스케줄러 구현
- 주기적 가격 정보 수집 스케줄러 추가
- DB에서 tracking 컬럼 true인 게임만 대상으로 가격 업데이트 수행
- tracking 컬럼 true가 유지되는 동안 PriceHistory db에 기록
- 무료 게임으로 바뀔시 tracking false로 설정

fix: tracking 컬럼 false 초기화 문제 수정

- 무료 게임 외 게임의 tracking 값이 잘못 false로 초기화되던 문제 해결
- 가격 0원/할인율 100% 처리 시 currentPrice 정상 반영

+++ src/main/java/com/example/steam_tracker/steam/service/SteamScrapingService.java
@@ -4,5 +4,5 @@ import java.util.Set;
 
 public interface SteamScrapingService {
     //start rank index 시작 등수 totalcount 총 추출할 게임ID 개수
-    Set<Integer> scrapingSteamAppId(int startRankIndex, int totalCount);
+    Set<Long> scrapingSteamAppId(int startRankIndex, int totalCount);
 }
